Move MetricsDataS3Client to dockstore-common
