+++ dockstore-common/src/main/java/io/dockstore/common/metrics/ObjectMetadata.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2023 OICR and UCSC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *           http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package io.dockstore.common.metrics;
+
+public enum ObjectMetadata {
+    OWNER("owner"),
+    DESCRIPTION("description");
+
+    private final String metadataKey;
+
+    ObjectMetadata(String metadata) {
+        this.metadataKey = metadata;
+    }
+
+    public String getMetadataKey() {
+        return metadataKey;
+    }
+
+    @Override
+    public String toString() {
+        return metadataKey;
+    }
+}
