+++ dockstore-common/src/test/java/io/dockstore/common/metrics/MetricsDataS3ClientTest.java
@@ -15,12 +15,11 @@
  *
  */
 
-package io.dockstore.webservice.core.metrics;
+package io.dockstore.common.metrics;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import io.dockstore.common.S3ClientHelper;
-import io.dockstore.common.metrics.MetricsData;
 import org.junit.jupiter.api.Test;
 
 class MetricsDataS3ClientTest {
