+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/ExtendedMetricsTRSOpenApiIT.java
@@ -54,6 +54,8 @@ import io.dockstore.common.Partner;
 import io.dockstore.common.Registry;
 import io.dockstore.common.SourceControl;
 import io.dockstore.common.metrics.MetricsData;
+import io.dockstore.common.metrics.MetricsDataMetadata;
+import io.dockstore.common.metrics.MetricsDataS3Client;
 import io.dockstore.openapi.client.ApiClient;
 import io.dockstore.openapi.client.ApiException;
 import io.dockstore.openapi.client.api.ContainersApi;
@@ -75,8 +77,6 @@ import io.dockstore.openapi.client.model.Workflow;
 import io.dockstore.openapi.client.model.WorkflowVersion;
 import io.dockstore.webservice.core.metrics.ExecutionTimeStatisticMetric;
 import io.dockstore.webservice.core.metrics.MemoryStatisticMetric;
-import io.dockstore.webservice.core.metrics.MetricsDataMetadata;
-import io.dockstore.webservice.core.metrics.MetricsDataS3Client;
 import java.io.IOException;
 import java.time.Instant;
 import java.util.ArrayList;
