--- dockstore-webservice/src/main/java/io/dockstore/webservice/core/metrics/MetricsDataMetadata.java
@@ -1,3 +1,0 @@
-package io.dockstore.webservice.core.metrics;
-
-public record MetricsDataMetadata(long owner, String description) {}
