--- dockstore-webservice/src/main/java/io/dockstore/webservice/core/metrics/ObjectMetadata.java
@@ -1,21 +1,0 @@
-package io.dockstore.webservice.core.metrics;
-
-public enum ObjectMetadata {
-    OWNER("owner"),
-    DESCRIPTION("description");
-
-    private final String metadataKey;
-
-    ObjectMetadata(String metadata) {
-        this.metadataKey = metadata;
-    }
-
-    public String getMetadataKey() {
-        return metadataKey;
-    }
-
-    @Override
-    public String toString() {
-        return metadataKey;
-    }
-}
